* Added fancy global-draft system, letting you just type during the night, and it records the text without having keyboard focus in any text box. (letting you then save it into an actual dream journal entry in the morning)

+++ Main/Source/LucidLink/Journal.tsx
@@ -3,11 +3,14 @@ import DialogAndroid from "react-native-dialogs";
 import Moment from "moment";
 
 import DreamUI from "./Journal/DreamUI";
-import {Assert, FromVDF, GetTypeName, ToVDF, E} from "../Frame/Globals";
+import {Assert, FromVDF, GetTypeName, ToVDF, E, Toast} from "../Frame/Globals";
 import {DatePickerAndroid, ScrollView, TouchableOpacity, Text} from "react-native";
 import Node from "../Packages/VTree/Node";
 import {LL} from "../LucidLink";
 import {P} from "../Packages/VDF/VDFTypeInfo";
+import {File} from "../Packages/V/VFile";
+import {autorun} from "mobx";
+import {observer} from "mobx-react/native";
 
 export class Dream extends Node {
 	static async Load(file) {
@@ -20,9 +23,20 @@ export class Dream extends Node {
 		return result;
 	}
 
-	Save() {
+	async Save() {
 		var vdf = ToVDF(this, false);
-		this.file.WriteAllText(vdf);
+		var oldFile = this.file;
+		this.file = this.NewFile;
+		await this.file.WriteAllText(vdf);
+
+		if (oldFile && oldFile.Path != this.file.Path)
+			oldFile.Delete();
+		this.fileOutdated = false;
+
+		if (this.Draft) {
+			LL.journal.draftDream = null;
+			LL.journal.loadedDreams.push(this);
+		}
 	}
 	Delete(onDone = null) {
 		var dialog = new DialogAndroid();
@@ -31,8 +45,12 @@ export class Dream extends Node {
 			"content": `Permanently delete dream?`,
 			"positiveText": "OK", "negativeText": "Cancel",
 			"onPositive": ()=> {
-				LL.journal.loadedDreams.Remove(this);
-				this.file.Delete();
+				if (this.Draft) {
+					LL.journal.draftDream = null;
+				} else {
+					LL.journal.loadedDreams.Remove(this);
+					this.file.Delete();
+				}
 				onDone && onDone();
 			},
 		});
@@ -46,21 +64,70 @@ export class Dream extends Node {
 
 		Assert(date instanceof Moment, `Date must be an instance of the Moment class, not ${GetTypeName(date)}.`);
 		this.date = date;
-		var journalFolder = LL.RootFolder.GetFolder("Journal");
-		this.file = journalFolder.GetFile(Moment().format("YYYY-MM-DD HH:mm:ss") + ".vdf");
+		this.file = this.NewFile;
 	}
 
-	file = null;
+	file: File = null; // can become outdated (until next save) if dream date is changed (or is draft dream)
+	get NewFile() {
+		var journalFolder = LL.RootFolder.GetFolder("Journal");
+		return journalFolder.GetFile(this.date.format("YYYY-MM-DD HH:mm:ss") + ".vdf");
+	}
+	@O fileOutdated = false;
 	date = null;
 
-	@P() name = null;
-	@P() text = null;
-	@P() lucid = false;
+	@P() @O name = null;
+	@P() @O text = null;
+	@P() @O lucid = false;
+
+	//@P() @O draft = false;
+	get Draft() {
+		return LL.journal.draftDream == this;
+	}
 }
 g.Extend({Dream});
 
 export class Journal extends Node {
-	loadedDreams = [];
+	constructor() {
+		super();
+		autorun(()=> {
+			// if we're loading draft-dream from last session, reset date to now (since info lost)
+			if (this.draftDream && this.draftDream.date == null)
+				this.draftDream.date = Moment();
+		});
+	}
+
+	@P() @O draftDream: Dream = null;
+	OnKeyDown(androidKeyCode: number, char: string) {
+		/*var isNumber = keyCode >= 7 && keyCode <= 16; // 0-9: code 7-16
+		var isLetter = keyCode >= 29 && keyCode <= 54; // a-z: code 29-54
+		//var isOtherTypingChar = [55, 56].Contains(keyCode);
+		var isOtherTypingChar = [",", ".", "'", "\"", "[", "]", "{", "}", "(", ")", "-", "_"].Contains(keyCode);
+		if (isNumber || isLetter || isOtherTypingChar) {*/
+
+		var specialChars = [24, 25, 4, 82]; // volume up, volume down, back, options
+		if (specialChars.Contains(androidKeyCode)) return;
+
+		var isBackspace = androidKeyCode == 67;
+		if (isBackspace && this.draftDream.text.length > 0) {
+			this.draftDream.text = this.draftDream.text.substr(0, this.draftDream.text.length - 1);
+			this.draftDream.fileOutdated = true;
+			return;
+		}
+
+		// if normal char (see: www.asciitable.com)
+		var charCode = char.charCodeAt(0);
+		var basicChar = charCode >= 32 && charCode <= 126;
+		var otherNormalChar = ["\n"].Contains(char);
+		if (basicChar || otherNormalChar) {
+			if (this.draftDream == null)
+				this.draftDream = new Dream(Moment()).Init({draft: true, text: ""});
+			this.draftDream.text += char;
+			this.draftDream.fileOutdated = true;
+			//Toast(`Text (${this.draftDream.text.length}, ${charCode}):` + this.draftDream.text);
+		}
+	}
+
+	loadedDreams: Dream[] = [];
 
 	async LoadDreamsForMonth(month, onFinish) {
 		var journalFolder = LL.RootFolder.GetFolder("Journal");
@@ -92,6 +159,7 @@ export class Journal extends Node {
 }
 g.Extend({Journal});
 
+@observer
 export class JournalUI extends Component<{} & BaseProps, {month?, openDream?}> {
 	state = {month: Moment(new Date().MonthDate), openDream: null};
 
@@ -113,14 +181,14 @@ export class JournalUI extends Component<{} & BaseProps, {month?, openDream?}> {
 		var node = LL.journal;
 
 		if (openDream) {
-			return <DreamUI dream={openDream} onBack={(save = true)=> {
+			return <DreamUI dream={openDream} onBack={()=> {
 				this.setState({openDream: null});
-				if (save)
-					openDream.Save();
 			}}/>
 		}
 
 		var entriesForMonth = LL.journal.GetLoadedDreamsForMonth(month);
+		if (node.draftDream)
+			entriesForMonth.push(node.draftDream);
 
 		return (
 			<Column>
@@ -157,15 +225,22 @@ export class JournalUI extends Component<{} & BaseProps, {month?, openDream?}> {
 	}
 }
 
-class DreamHeaderUI extends Component<{parent, dream, index, style?}, {}> {
+@observer
+class DreamHeaderUI extends Component<{parent, dream: Dream, index, style?}, {}> {
 	render() {
 		var {parent, dream, index, style} = this.props;
 		return (
 			<TouchableOpacity
 					style={E({backgroundColor: "#555", height: 100, borderRadius: 10, padding: 7}, index != 0 && {marginTop: 5})}
 					onPress={()=>parent.setState({openDream: dream})}>
-				<Text style={{fontSize: 18}}>{dream.name || "[no title]"}</Text>
-				<Text style={{marginTop: -25, fontSize: 18, textAlign: "right"}}>{dream.date.format("YYYY-MM-DD, HH:mm")}</Text>
+				<Row style={{marginTop: -3}}>
+					{dream.name && <Text style={{fontSize: 18, marginRight: 5}}>{dream.name}</Text>}
+					{dream.Draft && <Text style={{fontSize: 18, marginRight: 5}}>(draft)</Text>}
+					{dream.fileOutdated && <Text style={{fontSize: 18, color: "rgb(255,100,100)", marginRight: 5}}>(unsaved changes)</Text>}
+				</Row>
+				<Text style={{marginTop: dream.name || dream.Draft || dream.fileOutdated ? -25 : 0, fontSize: 18, textAlign: "right"}}>
+					{dream.date.format("YYYY-MM-DD, HH:mm")}
+				</Text>
 				{dream.lucid &&
 					<Text style={{position: "absolute", right: 7, fontSize: 18, textAlign: "right", color: "#0F0"}}>Lucid</Text>}
 				<Text>{dream.text}</Text>
